Always include settings.local.php
This is based on only a few minutes research, but I *think* the inclusion of settings.local.php and settings.platformsh.php should be reversed. AFAICT, the platform build process puts the automatic database settings in a settings.local.php, whereas this file was only including it when *not* on platform. When using the old version, I was landed in the Drupal installer. Using this change, everything seems to be working as it should.